# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ForgeCustomer API
  version: 0.1.0
  description: >
    Customer, commerce, licensing, entitlement, installation and usage authority for
    Boswell Digital Solutions products. Customer routes accept Supabase JWTs; admin routes
    require operator JWTs from a separate issuer. See docs/API.md.
servers:
  - url: https://api.forgecustomer.example
    description: production (placeholder)
...
paths:
...
  /v1/account:
    get:
      tags: [account]
      summary: Get own account summary
      responses:
        "200": { description: Account summary }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/account/provision:
    post:
      tags: [account]
      summary: Create or return the caller's ForgeCustomer profile
      description: >
        API-owned profile provisioning for a Supabase-authenticated user. The request may
        supply display/localization decoration only; customer type and commercial status
        are assigned by the server. Repeated calls are idempotent for the same auth user.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AccountProvisionRequest" }
      responses:
        "200":
          description: Provisioned or existing customer profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountProvisionResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { description: Profile validation failed }
...
  /v1/licenses:
    get:
      tags: [licensing]
      summary: Own licenses (with active device counts)
      responses:
        "200":
          description: License list
          content:
            application/json:
              schema:
                type: object
                required: [licenses]
                properties:
                  licenses:
                    type: array
                    items: { $ref: "#/components/schemas/License" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/installations:
    get:
      tags: [licensing]
      summary: List installations
      responses:
        "200":
          description: Installation list
          content:
            application/json:
              schema:
                type: object
                required: [installations]
                properties:
                  installations:
                    type: array
                    items: { $ref: "#/components/schemas/Installation" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [licensing]
      summary: Register an installation (idempotent by client-supplied install_key)
      description: >
        Registers an installed application instance and, when a device public key is
        supplied, the device identity it runs on. Repeat registration with the same
        install_key returns the existing installation; re-registering a deactivated
        installation reactivates the installation record (license slots are only consumed
        by activation). The client keeps its Ed25519 private key local and registers only
        the public key.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InstallationRegisterRequest" }
      responses:
        "200":
          description: Registered (or existing) installation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstallationRegisterResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Unknown or inactive product }
        "409": { description: Install key already registered for a different product }
        "422": { description: Registration validation failed }
  /v1/installations/{id}/activate:
    post:
      tags: [licensing]
      summary: Activate a license on an installation (enforces device limit)
      description: >
        Links the customer's license to the installation. When license_id is omitted the
        most recently issued active license for the installation's product is used.
        Fails closed on suspended/expired/revoked licenses and on explicit revocation
        records; concurrent activations are serialized per license. Idempotent for an
        already-active pairing.
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                license_id:
                  type: ["string", "null"]
                  format: uuid
                  description: Explicit license to activate; defaults to the product's active license.
      responses:
        "200":
          description: Activation result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstallationActivateResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402": { description: "Device limit reached (DEVICE_LIMIT_REACHED)" }
        "403": { description: "License not active, or activation blocked by revocation (REVOKED)" }
        "404": { description: "Installation or license not found" }
        "409": { description: "Installation is deactivated; register it again first" }
  /v1/installations/{id}/heartbeat:
    post:
      tags: [licensing]
      summary: Installation heartbeat (liveness; may refresh reported app version)
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                app_version: { type: ["string", "null"], maxLength: 64 }
      responses:
        "200":
          description: Heartbeat recorded
          content:
            application/json:
              schema:
                type: object
                required: [installation_id, status]
                properties:
                  installation_id: { type: string, format: uuid }
                  status: { type: string, enum: [active, deactivated] }
                  app_version: { type: ["string", "null"] }
                  last_heartbeat_at: { type: ["string", "null"], format: date-time }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Installation not found }
  /v1/installations/{id}/deactivate:
    post:
      tags: [licensing]
      summary: Deactivate an installation (releases its license activations)
      description: >
        Marks the installation deactivated and releases its active license activations,
        freeing device slots. Idempotent.
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200":
          description: Deactivation result
          content:
            application/json:
              schema:
                type: object
                required: [installation_id, status, released_activations, already_deactivated]
                properties:
                  installation_id: { type: string, format: uuid }
                  status: { type: string, enum: [deactivated] }
                  released_activations: { type: integer, minimum: 0 }
                  already_deactivated: { type: boolean }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Installation not found }
  /v1/devices:
    get:
      tags: [licensing]
      summary: Own devices
      responses:
        "200":
          description: Device list
          content:
            application/json:
              schema:
                type: object
                required: [devices]
                properties:
                  devices:
                    type: array
                    items: { $ref: "#/components/schemas/Device" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/entitlements/current:
    get:
      tags: [entitlements]
      summary: Current signed entitlement snapshot
      description: >
        Assembles the caller's entitlements from the included-plan baseline, current
        subscription plan, license grants, promotional grants, and admin overrides, with
        cloud features gated by subscription truth and denials failing closed. The
        snapshot is Ed25519-signed and stored for audit/replay before being returned.
        The wire JSON field order matches the canonical signing order.
      parameters:
        - name: product_key
          in: query
          required: false
          schema: { type: string, default: authorforge }
        - name: installation_id
          in: query
          required: false
          description: Bind the snapshot to an owned installation (overrides product_key).
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Signed snapshot
          content:
            application/json:
              schema: { $ref: "./entitlement-v1.schema.json" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: "Unknown product, or installation not found" }
        "422": { description: Invalid product key or installation id }
  /v1/entitlements/check:
    post:
      tags: [entitlements]
      summary: Check a specific feature or quota (advisory, read-only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EntitlementCheckRequest" }
      responses:
        "200":
          description: Check result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EntitlementCheckResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Unknown or inactive product }
        "422": { description: "Provide exactly one valid feature_key or quota_key" }
  /v1/entitlements/offline-lease:
    post:
      tags: [entitlements]
      summary: Issue a signed offline lease for an activated installation
      description: >
        Fails closed: requires an active installation holding an active activation on an
        active, unexpired license, with no explicit revocation in scope and a non-revoked
        device. Lease lifetime is the configured offline grace window. Each issued lease
        is stored and audited.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [installation_id]
              properties:
                installation_id: { type: string, format: uuid }
      responses:
        "200":
          description: Signed offline lease (wire field order matches signing order)
          content:
            application/json:
              schema: { $ref: "./lease-v1.schema.json" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: "Suspended customer, non-active license, or revocation (REVOKED) — no new lease" }
        "404": { description: Installation not found }
        "409": { description: "Installation deactivated, or no active activation" }
...
  /v1/usage/reserve:
    post:
      tags: [usage]
      summary: Reserve usage
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses: { "401": { $ref: "#/components/responses/Unauthorized" }, "402": { description: Quota exceeded } }
  /v1/usage/commit:
    post:
      tags: [usage]
      summary: Commit reserved/direct usage
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses: { "401": { $ref: "#/components/responses/Unauthorized" } }
...
  /v1/webhooks/stripe:
    post:
      tags: [webhooks]
      summary: Stripe webhook receiver (signature-verified; idempotent processing)
      security: []
      responses:
        "200":
          description: Acknowledged and stored or deduped
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StripeWebhookResponse" }
        "400": { description: Invalid signature }
...
components:
  securitySchemes:
    customerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase-issued customer JWT
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Operator JWT from the admin issuer (separate from customer auth)
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not permitted (or suspended)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    AccountProvisionRequest:
      type: object
      additionalProperties: false
      properties:
        display_name: { type: ["string", "null"], maxLength: 120 }
        country_code:
          type: ["string", "null"]
          description: Two-letter ISO country code; normalized to uppercase.
        timezone: { type: ["string", "null"], maxLength: 80 }
...
security:
  - customerAuth: []